chore: turn store filter into general issue filter
