+++ packages/app/lib/apps/issue-filter/index.js
@@ -1,0 +1,5 @@
+import IssueFilter from './IssueFilter.js';
+
+export default {
+  issueFilter: [ 'type', IssueFilter ]
+};
